+++ Runtime/Ponito/Core/Asyncs/CoroutineResult.cs
@@ -1,0 +1,13 @@
+namespace Ponito.Core.Asyncs
+{
+    /// <summary>
+    ///     封裝協程的結果
+    /// </summary>
+    public class CoroutineResult<T>
+    {
+        /// <summary>
+        ///     協程的返回值
+        /// </summary>
+        public T value;
+    }
+}