+++ Runtime/Ponito/Core/Tests/ForTest.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Ponito.Core.Tests
+{
+    [AttributeUsage(AttributeTargets.Field)]
+    public class ForTest : Attribute
+    {
+        
+    }
+}